added packages and application.properties
