--- src/main/java/com/happykealifornia/hk/Domain/Customer.java
@@ -1,4 +1,0 @@
-package com.happykealifornia.hk.Domain;
-
-public class Customer {
-}
