--- src/main/java/com/happykealifornia/hk/Domain/Hotel.java
@@ -1,4 +1,0 @@
-package com.happykealifornia.hk.Domain;
-
-public class Hotel {
-}
