--- src/main/java/com/happykealifornia/hk/Domain/Payment.java
@@ -1,4 +1,0 @@
-package com.happykealifornia.hk.Domain;
-
-public class Payment {
-}
