+++ src/main/java/com/happykealifornia/hk/Model/Booking.java
@@ -1,0 +1,4 @@
+package com.happykealifornia.hk.Model;
+
+public class Booking {
+}
