+++ src/main/java/com/happykealifornia/hk/Model/Customer.java
@@ -1,0 +1,4 @@
+package com.happykealifornia.hk.Model;
+
+public class Customer {
+}
