+++ src/main/java/com/happykealifornia/hk/Model/Hotel.java
@@ -1,0 +1,4 @@
+package com.happykealifornia.hk.Model;
+
+public class Hotel {
+}
